Second sha256.c optimization
Achieve a significant performance increase by unrolling all SHA-256
loops and allowing constant expressions to be simplified at compile
time.

Another optimization was attempted in which the SHA-256 variables
were rotated in place with each call to the ROUND() macro, potentially
eliminating many register moves. However, the compiler appears to be
smart enough to do this itself, so the optimization was not kept as
it makes the code much more difficult to read.