Isolate getwork/sendwork into separate threads; cleanup
Instead of trying to both poll for work and submit solutions in the same
thread, do these in separate threads. If a submitted solution is not
accepted, the polling thread is run immediately to get new work, under
the assumption a new block is being worked on.

We now report the number of backlogged work items along with the hashing
rate. We might need to consider trimming these occasionally, not just
each time a new block is being worked on.