Implement long polling
Support long polling by creating a new RPCProxy object for the long poll
URI and yielding it to the block attached to the current RPC call.

Split the getwork thread into a gathering thread and a processing thread
so that long polling can execute in a subthread of the gathering thread.

We currently use a long polling timeout of 12 hours.