First pass persistent HTTP clean-up
The timeout is only relevant to long polling, so don't set the open
timeout with it. String#to_uri doesn't seem to exist, so revert back
to URI.parse. We don't need quite so much header flexibility, so just
modify the User-Agent header directly on the request.

While we're cleaning things up, remove external access to the RPCProxy
attributes (until we have a compelling need for them).